Category

+++ src/main/java/com/example/ddashmanagement/Services/Impl/CategoryService.java
@@ -1,5 +1,7 @@
 package com.example.ddashmanagement.Services.Impl;
 
+import com.example.ddashmanagement.Ennum.EtatCategory;
+import com.example.ddashmanagement.Ennum.StatusCategorie;
 import com.example.ddashmanagement.Ennum.TypeCategory;
 import com.example.ddashmanagement.Entites.Category;
 import com.example.ddashmanagement.Entites.CategoryFille;
@@ -8,6 +10,7 @@ import com.example.ddashmanagement.Repository.CategoryRepository;
 import com.example.ddashmanagement.Services.IServiceCategory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -36,9 +39,18 @@ public class CategoryService implements IServiceCategory {
     }
 
     @Override
-    public Category updateCategorie(Category c) {
-        return null;
+    public CategoryFille updateCategorie(String id, CategoryFille c) {
+
+        CategoryFille cat = categoryFilleRepository.findById(id).orElseThrow(() -> new RuntimeException("Entité non trouvée avec id " + id));
+
+        if (cat.getProducts().isEmpty()) {
+            cat.setLibeléCategorie(c.getLibeléCategorie());
+            cat.setCritere(c.getCritere());
+        }
+
+        return categoryFilleRepository.save(cat);
     }
+
 
     @Override
     public List<CategoryFille> findAllCategorie() {
@@ -46,13 +58,26 @@ public class CategoryService implements IServiceCategory {
     }
 
     @Override
-    public void deleteCategorie(Category c) {
-        categoryFilleRepository.delete(c);
-
+    public void deleteCategorie(CategoryFille c) {
+        if (c.getProducts().isEmpty()){
+            categoryFilleRepository.delete(c);
+        }
     }
 
     @Override
     public boolean categorieExists(String id) {
         return categoryFilleRepository.existsById(id);
     }
+
+    @Override
+    public void ChangerStatus(CategoryFille c) {
+
+    }
+
+    @Override
+    public List<CategoryFille> Filtre(StatusCategorie status, TypeCategory type, EtatCategory etat) {
+        return categoryFilleRepository.findCategoryFilleByEtatOrStatusOrType(status , type , etat);
+    }
+
+
 }
